+++ src/app/Helpers/form-validators.ts
@@ -1,0 +1,94 @@
+import {
+    ReactiveFormsModule,
+    NG_VALIDATORS,
+    FormsModule,
+    FormGroup,
+    FormControl,
+    ValidatorFn,
+    Validator
+  } from '@angular/forms';
+  import { Directive } from '@angular/core';
+
+  @Directive({
+    selector: '[emailvalidator][ngModel]',
+    providers: [
+      {
+        provide: NG_VALIDATORS,
+        useExisting: EmailValidator,
+        multi: true
+      }
+    ]
+  })
+ 
+
+export class EmailValidator implements Validator {
+    validator: ValidatorFn;
+    
+    constructor() {
+      this.validator = emailValidator();
+    }
+    
+    validate(c: FormControl) {
+      return this.validator(c);
+    }
+  }
+
+  function emailValidator() : ValidatorFn {
+    return (c: FormControl) => {
+      
+      let isValid = /^[_a-z0-9]+(\.[_a-z0-9]+)*@[a-z0-9-]+(\.[a-z0-9-]+)*(\.[a-z]{2,4})$/.test(c.value);
+      
+      if(isValid) {
+        return null;
+      } else {
+        return {
+          emailvalidator : {
+            valid: false
+          }
+        };
+      }
+    }
+   }  
+
+ 
+  @Directive({
+    selector: '[nicvalidator][ngModel]',
+    providers: [
+      {
+        provide: NG_VALIDATORS,
+        useExisting: NicValidator,
+        multi: true
+      }
+    ]
+  })
+
+ export class NicValidator implements Validator {
+    validator: ValidatorFn;
+    
+    constructor() {
+      this.validator = nicValidator();
+    }
+    
+    validate(c: FormControl) {
+      return this.validator(c);
+    }
+  }   
+  
+  function nicValidator() : ValidatorFn {
+    return (c: FormControl) => {
+      
+      let isValid = /^[0-9]{9}[vVxX]$/.test(c.value);
+      
+      if(isValid) {
+        return null;
+      } else {
+        return {
+          nicvalidator : {
+            valid: false
+          }
+        };
+      }
+    }
+   }
+
+
