+++ src/app/components/layout/user-page/user-page.module.ts
@@ -6,8 +6,8 @@ import { UserPageRoutingModule } from './user-page-routing.module';
 
 import { UserPageComponent } from './user-page.component';
 import { UserListComponent } from './user-list/user-list.component';
-import { UserSaveFormComponent , EmailValidator } from './user-save-form/user-save-form.component';
-
+import { UserSaveFormComponent } from './user-save-form/user-save-form.component';
+import { EmailValidator , NicValidator} from './../../../Helpers/form-validators';
 import { UserService } from './../../../services/user.service';
 
 @NgModule({
@@ -21,7 +21,8 @@ import { UserService } from './../../../services/user.service';
     UserPageComponent,
     UserListComponent,
     UserSaveFormComponent,
-    EmailValidator
+    EmailValidator,
+    NicValidator
   ],
   providers: [
     UserService
