+++ src/app/components/layout/user-page/user-save-form/user-save-form.component.ts
@@ -1,13 +1,4 @@
 import { Component, OnInit, ViewChild, Directive } from '@angular/core';
-import {
-  ReactiveFormsModule,
-  NG_VALIDATORS,
-  FormsModule,
-  FormGroup,
-  FormControl,
-  ValidatorFn,
-  Validator
-} from '@angular/forms';
 import { UserService } from './../../../../services/user.service';
 import { User } from './../../../../models/user';
 
@@ -36,43 +27,4 @@ export class UserSaveFormComponent implements OnInit {
 
 }
 
-@Directive({
-  selector: '[emailvalidator][ngModel]',
-  providers: [
-    {
-      provide: NG_VALIDATORS,
-      useExisting: EmailValidator,
-      multi: true
-    }
-  ]
-})
 
-export class EmailValidator implements Validator {
-  validator: ValidatorFn;
-  
-  constructor() {
-    this.validator = emailValidator();
-  }
-  
-  validate(c: FormControl) {
-    return this.validator(c);
-  }
-}
-
-function emailValidator() : ValidatorFn {
-  return (c: FormControl) => {
-    
-    let isValid = /^[_a-z0-9]+(\.[_a-z0-9]+)*@[a-z0-9-]+(\.[a-z0-9-]+)*(\.[a-z]{2,4})$/.test(c.value);
-    
-    if(isValid) {
-      return null;
-    } else {
-      return {
-        emailvalidator : {
-          valid: false
-        }
-      };
-    }
-
-  }
-}
